% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnTrain.R
\name{.nnWeightBase}
\alias{.nnWeightBase}
\title{Compute the par_ptr base of a network's weight block for a model}
\usage{
.nnWeightBase(model, nnId, K, H)
}
\description{
Compute the par_ptr base of a network's weight block for a model
}
\keyword{internal}
