% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnJoint.R
\name{getValidNlmixrCtl.nn}
\alias{getValidNlmixrCtl.nn}
\title{Validate the control for `est = "nn"`}
\usage{
\method{getValidNlmixrCtl}{nn}(control)
}
\arguments{
\item{control}{the control passed to `nlmixr2()` (as a length-1 list).}
}
\value{
a valid `nnControl` object.
}
\description{
Validate the control for `est = "nn"`
}
