% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnControl.R
\name{getValidNlmixrCtl.nnIter}
\alias{getValidNlmixrCtl.nnIter}
\title{Validate the control for `est = "nnIter"`}
\usage{
\method{getValidNlmixrCtl}{nnIter}(control)
}
\arguments{
\item{control}{the control passed to `nlmixr2()` (as a length-1 list).}
}
\value{
a valid `nnIterControl` object.
}
\description{
Validate the control for `est = "nnIter"`
}
