% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnTorch.R
\name{nnTorchForwardBatch}
\alias{nnTorchForwardBatch}
\title{Batch forward pass of a torch module}
\usage{
nnTorchForwardBatch(id, X)
}
\arguments{
\item{id}{network id.}

\item{X}{numeric matrix of inputs, `N` rows by `K` columns.}
}
\value{
numeric vector of length `N`.
}
\description{
Batch forward pass of a torch module
}
\keyword{internal}
