% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnTorch.R
\name{nnTorchOptInit}
\alias{nnTorchOptInit}
\alias{nnTorchZeroGrad}
\title{Create an optimizer for a torch module}
\usage{
nnTorchOptInit(id, type = "adam", lr = 0.05)

nnTorchZeroGrad(id)
}
\arguments{
\item{id}{network id.}

\item{type}{"adam" or "sgd".}

\item{lr}{learning rate.}
}
\value{
invisibly NULL.
}
\description{
Create an optimizer for a torch module
}
