% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnTorch.R
\name{nnTorchSetWeights}
\alias{nnTorchSetWeights}
\title{Load flat weights into a torch module}
\usage{
nnTorchSetWeights(id, values)
}
\arguments{
\item{id}{network id.}

\item{values}{numeric weight vector.}
}
\value{
invisibly NULL.
}
\description{
Load flat weights into a torch module
}
\keyword{internal}
