% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnTorch.R
\name{nnTorchStep}
\alias{nnTorchStep}
\title{Optimizer step; also syncs updated weights into the loader buffer}
\usage{
nnTorchStep(id)
}
\arguments{
\item{id}{network id.}
}
\value{
invisibly NULL.
}
\description{
Optimizer step; also syncs updated weights into the loader buffer
}
\keyword{internal}
