% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnTorch.R
\name{nnTorchWeights}
\alias{nnTorchWeights}
\title{Flat weights of a torch module (nnWeightLayout order)}
\usage{
nnTorchWeights(id)
}
\arguments{
\item{id}{network id.}
}
\value{
numeric vector.
}
\description{
Flat weights of a torch module (nnWeightLayout order)
}
\keyword{internal}
