% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr2nn-package.R
\docType{package}
\name{nlmixr2nn-package}
\alias{nlmixr2nn}
\alias{nlmixr2nn-package}
\title{nlmixr2nn: Neural-Network ODEs for 'rxode2' and 'nlmixr2'}
\description{
Embed neural networks inside ODE right-hand sides for rxode2/nlmixr2, backed
by a C++ libtorch module whose weights are injected into the solve parameter
vector by an rxode2 par-loader hook.
}
\author{
\strong{Maintainer}: Matthew Fidler \email{matthew.fidler@gmail.com}

Authors:
\itemize{
  \item Matthew Fidler \email{matthew.fidler@gmail.com}
}

}
\keyword{internal}
