% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnTorch.R
\name{nnTorchForward}
\alias{nnTorchForward}
\title{Forward pass of a torch module at one input (for validation)}
\usage{
nnTorchForward(id, x)
}
\arguments{
\item{id}{network id.}

\item{x}{numeric input vector of length K.}
}
\value{
numeric scalar.
}
\description{
Forward pass of a torch module at one input (for validation)
}
\keyword{internal}
