% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnTorch.R
\name{nnTorchGetGrad}
\alias{nnTorchGetGrad}
\title{Flattened parameter gradients (nnWeightLayout order)}
\usage{
nnTorchGetGrad(id)
}
\arguments{
\item{id}{network id.}
}
\value{
numeric vector.
}
\description{
Flattened parameter gradients (nnWeightLayout order)
}
\keyword{internal}
