% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnTorch.R
\name{nnTorchSave}
\alias{nnTorchSave}
\alias{nnTorchLoad}
\title{Save / load a trained torch module to a binary}
\usage{
nnTorchSave(id, file)

nnTorchLoad(id, file)
}
\arguments{
\item{id}{network id.}

\item{file}{path to the serialized module.}
}
\value{
invisibly NULL.
}
\description{
Save / load a trained torch module to a binary
}
\keyword{internal}
